Add fish task
Score: 62%, Correctness: 100%.
For every fish we check all other fishes on the side it swims. If the second is bigger and swims the same direction, we start considering its size, because it can rescue the first from opposite fishes on its(second's) side. If the second is larger and swims opposite it eats the first, so we count it.